requirements.txt
anyio==3.5.0
asgiref==3.4.1
click==8.0.3
fastapi==0.72.0
h11==0.12.0
httptools==0.3.0
idna==3.3
numpy==1.22.1
pandas==1.3.5
peewee==3.14.8
psycopg2-binary==2.9.3
pydantic==1.9.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
scipy==1.7.3
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.0.1
uvicorn==0.17.0
uvloop==0.16.0
watchgod==0.7
websockets==10.1
